test(testdata): rename & restyle the benchmark
