+++ testdata/kitex_gen/pb/baseline/baseline_test_utils.go
@@ -1,0 +1,121 @@
+package baseline
+
+func (x *Nesting) SizeField1() (n int) {
+	return x.sizeField1()
+}
+
+func (x *Nesting) SizeField2() (n int) {
+	return x.sizeField2()
+}
+
+func (x *Nesting) SizeField3() (n int) {
+	return x.sizeField3()
+}
+
+func (x *Nesting) SizeField4() (n int) {
+	return x.sizeField4()
+}
+
+func (x *Nesting) SizeField5() (n int) {
+	return x.sizeField5()
+}
+
+func (x *Nesting) SizeField6() (n int) {
+	return x.sizeField6()
+}
+
+func (x *Nesting) SizeField7() (n int) {
+	return x.sizeField7()
+}
+
+func (x *Nesting) SizeField8() (n int) {
+	return x.sizeField8()
+}
+
+func (x *Nesting) SizeField9() (n int) {
+	return x.sizeField9()
+}
+
+func (x *Nesting) SizeField10() (n int) {
+	return x.sizeField10()
+}
+
+func (x *Nesting) SizeField11() (n int) {
+	return x.sizeField11()
+}
+
+func (x *Nesting) SizeField12() (n int) {
+	return x.sizeField12()
+}
+
+func (x *Nesting) SizeField13() (n int) {
+	return x.sizeField13()
+}
+
+func (x *Nesting) SizeField14() (n int) {
+	return x.sizeField14()
+}
+
+func (x *Nesting) SizeField15() (n int) {
+	return x.sizeField15()
+}
+
+func (x *Nesting) FastWriteField1(buf []byte) (offset int) {
+	return x.fastWriteField1(buf)
+}
+
+func (x *Nesting) FastWriteField2(buf []byte) (offset int) {
+	return x.fastWriteField2(buf)
+}
+
+func (x *Nesting) FastWriteField3(buf []byte) (offset int) {
+	return x.fastWriteField3(buf)
+}
+
+func (x *Nesting) FastWriteField4(buf []byte) (offset int) {
+	return x.fastWriteField4(buf)
+}
+
+func (x *Nesting) FastWriteField5(buf []byte) (offset int) {
+	return x.fastWriteField5(buf)
+}
+
+func (x *Nesting) FastWriteField6(buf []byte) (offset int) {
+	return x.fastWriteField6(buf)
+}
+
+func (x *Nesting) FastWriteField7(buf []byte) (offset int) {
+	return x.fastWriteField7(buf)
+}
+
+func (x *Nesting) FastWriteField8(buf []byte) (offset int) {
+	return x.fastWriteField8(buf)
+}
+
+func (x *Nesting) FastWriteField9(buf []byte) (offset int) {
+	return x.fastWriteField9(buf)
+}
+
+func (x *Nesting) FastWriteField10(buf []byte) (offset int) {
+	return x.fastWriteField10(buf)
+}
+
+func (x *Nesting) FastWriteField11(buf []byte) (offset int) {
+	return x.fastWriteField11(buf)
+}
+
+func (x *Nesting) FastWriteField12(buf []byte) (offset int) {
+	return x.fastWriteField12(buf)
+}
+
+func (x *Nesting) FastWriteField13(buf []byte) (offset int) {
+	return x.fastWriteField13(buf)
+}
+
+func (x *Nesting) FastWriteField14(buf []byte) (offset int) {
+	return x.fastWriteField14(buf)
+}
+
+func (x *Nesting) FastWriteField15(buf []byte) (offset int) {
+	return x.fastWriteField15(buf)
+}
